requirements.txt
boto3==1.14.7
discord.py==1.3.3
